Can you modify your API to support the borders like so?
app.UploadImage(imgurl, borderSize, borderColor);

borderSize: integer - pixel width of border
borderColor: string - HTML hex color (e.g. #FFFFFF)

If borderSize=0, then use the image as-is (what we have currently)
If borderSize > 0 then shrink the image width and height by 2x the borderSize and then add the border in the color specified. It's important not to cover the image with the border, but rather to shrink the image first to make room for the border.

-done